dhc2 v2.4.0: replace the "copilot" with the default autopilot.
The Autopilot | Autopilot Settings menu entry was hijacked to open a
custom "copilot" dialog instead of the default Autopilot Settings
dialog.  When the toggle button "Copilot has control", a basic and
custom autopilot was enabled that did wing leveling and vertical speed
hold with VS=0.

Bug: the default Autopilot Settings dialog was still available with
F11 but was inoperative.
Bug: the "copilot" would constantly drift to the right.
Bug: the "copilot" would not maintain VS=0 but would gain altitude
slowly.

To fix these problems, I deleted the "copilot" dialog and custom
autopilot; the default autopilot in 2019.1.1 happens to work well
enough (except in "speed with pitch" mode).