Harrier-GR3: various improvements to auto-hover.
Use property nodes more consistently to reduce overheads.

Made enabling/disabling of the various auto-hover modes, and selection of
PID/Nasal implementations more systematic.

Added PID controllers for forwards and sideways accelaration.

Improved enabling/disabling when leaving/touching ground. We now also activate
pid controllers if no wheels on ground, or number of wheels on ground has
decreased, which helps on takeoff particularly if controls are not centred.

Alt-# toggles use of PID controllers for auto-hover. Default is off - PID
controllers are improved but still not as good as the Nasal functions.