Aircraft/MPCarrier/Systems/MPCarriers.nas: added optional enabling of ai scenarios plus improved attach.
If /sim/mp-carriers/auto-attach is true, we enable the required ai scenerio
when we see a multiplayer carrier (and open a popup to notify the user), and
also attach more eagerly.

Always latch ai carrier to mutiplayer carrier if we are replaying; this appears
to improve handling of replay if multiplayer replay is enabled.

Added some extra info to /ai/models/carrier[]/mp-control/.